+++ CHANGELOG.md
@@ -1,11 +1,19 @@
 # Changelog
 
-## Unreleased
+## 23.11.2
 
 **Features**:
 
 - `normalize_performance_score` now handles `PerformanceScoreProfile` configs with zero weight components and component weight sums of any number greater than 0. ([#2756](https://github.com/getsentry/relay/pull/2756))
 - Return global config ready status to downstream relays.  ([#2765](https://github.com/getsentry/relay/pull/2765))
+
+**Features**:
+
+- Normalize invalid metric names. ([#2769](https://github.com/getsentry/relay/pull/2769))
+
+**Internal**:
+
+- Add support for metric metadata. ([#2751](https://github.com/getsentry/relay/pull/2751))
 
 ## 23.11.1